Removed `Node`
