+++ src/rule.rs
@@ -209,3 +209,52 @@ impl Rule {
         }
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::super::*;
+    use std::rc::Rc;
+    use std::cell::RefCell;
+
+    #[test]
+    fn node_ref() {
+        // Create a node rule the refers to itself.
+        let foo: Rc<String> = Rc::new("foo".into());
+        let num: Rc<String> = Rc::new("num".into());
+        let node = Rc::new(RefCell::new(Rule::Sequence(Sequence {
+            debug_id: 1,
+            args: vec![
+                Rule::Number(Number {
+                    debug_id: 2,
+                    property: Some(num.clone()),
+                    allow_underscore: false,
+                }),
+                Rule::Optional(Box::new(Optional {
+                    debug_id: 3,
+                    rule: Rule::Sequence(Sequence {
+                        debug_id: 4,
+                        args: vec![
+                            Rule::Whitespace(Whitespace {
+                                debug_id: 3,
+                                optional: false
+                            }),
+                            Rule::Node(NodeRef::Name(foo.clone(), 3)),
+                        ]
+                    }),
+                })),
+            ],
+        })));
+
+        // Replace self referencing names with direct references.
+        let refs = vec![(foo.clone(), node.clone())];
+        let mut rules = Rule::Node(NodeRef::Name(foo.clone(), 0));
+        rules.update_refs(&refs);
+
+        let text = "1 2 3";
+        let data = parse(&rules, &refs, text).unwrap();
+        assert_eq!(data.len(), 3);
+        assert_eq!(&data[0].1, &MetaData::F64(num.clone(), 1.0));
+        assert_eq!(&data[1].1, &MetaData::F64(num.clone(), 2.0));
+        assert_eq!(&data[2].1, &MetaData::F64(num.clone(), 3.0));
+    }
+}
